Stubbing out reworked Debug Logger
Wanted to make this a little smarter and group log messages by request. Thought there is that using the request ID we can group all log messages that came in while the page was rendering.

This will also allow us to support AJAX requests if they get fired off after page load then we can check for those new messages.

Currently using the Request ID as the key as this will allow for rendering in Vue and not break possible state changes if an AJAX event is fired and new list is returned ( index for selected request should remain unchanged )

TODO: Figure out a better way to prune old messages.  Right now this is on a timer, but probably not ideal.  It would be best to figure out when a new page load even is happening and trigger a clean on this data... but we need to make sure it does NOT happen for AJAX calls that could happen after page load.  We may also use controllers to load that debugger data outside the render process, so we can't "just wipe" the data with every.

Current solution is to just log messages for 30 seconds and auto prune with each page request.  May need to make that longer, may need to come up with an entire new solution.  We'll see, ha ha