Got AJAX Calls Working for Console :D
Also updated the way the debug messages are working, as technically each unique request should rebuild the debugger, UNLESS it was an ajax call after page load, so really just needed to see if the request was made via ajax.  If not, start the debugger from scratch, otherwise the page is loaded and all future requests on the page will be AJAX, which we want to collect as part of this session.